-reorganized src/app folder by adding more subfolders and shuffling files around. -updated package.json to include v23 contributors -accepted a merge conflict in typedoc.json
