+++ .gitignore
@@ -13,6 +13,5 @@ src/extension/build.pem
 bower_components
 sandboxes/manual-tests/NextJS/.next
 .vscode
-src/app/components/Map.tsx
 package-lock.json
 yarn.lock
