+++ package.json
@@ -55,6 +55,7 @@
     "David Bernstein",
     "David Chai",
     "David Kim",
+    "David Moore",
     "Dennis Lopez",
     "Edar Liu",
     "Edwin Menendez",
@@ -73,7 +74,9 @@
     "James Nghiem",
     "Jasmine Noor",
     "Jason Victor",
+    "Jesse Rosengrant",
     "Jimmy Phy",
+    "John Banks",
     "Joseph Park",
     "Joseph Stern",
     "Josh Kim",
@@ -87,6 +90,7 @@
     "Kristina Wallen",
     "Kyle Bell",
     "Lance Ziegler",
+    "Liam Donaher",
     "Lina Shin",
     "Mark Teets",
     "Mike Bednarz",
