+++ src/app/__tests__/ActionContainer.test.tsx
@@ -132,7 +132,7 @@ const customStore = configureStore({
 const render = (component) => rtlRender(<Provider store={customStore}>{component}</Provider>);
 
 const MockRouteDescription = jest.fn();
-jest.mock('../components/RouteDescription', () => () => {
+jest.mock('../components/Actions/RouteDescription', () => () => {
   MockRouteDescription();
   return <div>MockRouteDescription</div>;
 });
