+++ src/app/__tests__/MainSlider.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render as rtlRender, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import MainSlider from '../components/MainSlider';
+import MainSlider from '../components/TimeTravel/MainSlider';
 import { mainSlice } from '../slices/mainSlice';
 import { Provider } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
