+++ src/app/__tests__/TravelContainer.test.tsx
@@ -6,7 +6,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import { mainSlice } from '../slices/mainSlice';
 import { useDispatch } from 'react-redux';
 import { ThemeProvider } from '@mui/material/styles';
-import theme from '../components/theme';
+import theme from '../styles/theme';
 import '@testing-library/jest-dom'; // needed this to extend the jest-dom assertions  (ex toHaveTextContent)
 
 const customTabs = {
