+++ src/app/components/Buttons/Tutorial.tsx
@@ -5,11 +5,11 @@
 import * as React from 'react';
 import { Component } from 'react';
 import 'intro.js/introjs.css';
-import { TutorialProps, TutorialState, StepsObj } from '../FrontendTypes';
+import { TutorialProps, TutorialState, StepsObj } from '../../FrontendTypes';
 import { Button } from '@mui/material';
 import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
 const { Steps } = require('intro.js-react'); //Must be required in. This enables compatibility with TS. If imported in, throws ts error of not rendering steps as a class component correctly. The package 'intro.js-react' is small React wrapper around Intro.js. The wrapper provides support for both steps and hints. https://introjs.com/docs/
-import { setCurrentTabInApp, tutorialSaveSeriesToggle } from '../slices/mainSlice';
+import { setCurrentTabInApp, tutorialSaveSeriesToggle } from '../../slices/mainSlice';
 import { useDispatch, useSelector } from 'react-redux';
 /*
   This is the tutorial displayed when the "How to use" button is clicked
@@ -365,7 +365,7 @@ export default class Tutorial extends Component<TutorialProps, TutorialState> {
           type='button'
           onClick={() => startIntro()}
         >
-          <HelpOutlineIcon className="button-icon" sx={{ pr: 1 }} /> Tutorial
+          <HelpOutlineIcon className='button-icon' sx={{ pr: 1 }} /> Tutorial
         </Button>
       </>
     );
