+++ src/app/components/DiffRoute/Diff.tsx
@@ -1,8 +1,14 @@
 import React from 'react';
-import { diff, formatters} from 'jsondiffpatch';
+import { diff, formatters } from 'jsondiffpatch';
 const jsondiffpatch = require('jsondiffpatch');
 import ReactHtmlParser from 'html-react-parser';
-import { CurrentTab, DiffProps, MainState, RootState, StatelessCleaning } from '../FrontendTypes';
+import {
+  CurrentTab,
+  DiffProps,
+  MainState,
+  RootState,
+  StatelessCleaning,
+} from '../../FrontendTypes';
 import { useSelector } from 'react-redux';
 
 /**
