+++ src/app/components/StateRoute/WebMetrics/WebMetrics.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import Charts from 'react-apexcharts';
 import ReactHover, { Trigger, Hover } from 'react-hover';
-import { OptionsCursorTrueWithMargin } from '../FrontendTypes';
-import { setCurrentTabInApp } from '../slices/mainSlice';
+import { OptionsCursorTrueWithMargin } from '../../../FrontendTypes';
+import { setCurrentTabInApp } from '../../../slices/mainSlice';
 import { useDispatch } from 'react-redux';
 /*
   Used to render a single radial graph on the 'Web Metrics' tab
