+++ src/app/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/jsx-filename-extension */
 import React from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './components/App';
+import App from './App';
 import './styles/main.scss';
 import { store } from './store'; //imported RTK Store
 import { Provider } from 'react-redux'; //imported Provider
