--- src/app/jsondiffpatch.d.ts
@@ -1,3 +1,0 @@
-declare module 'jsondiffpatch/formatters/html' {
-  export * as formatters from 'jsondiffpatch/lib/formatters/html';
-}
